Refatorado RESTAPI para seu próprio pacote
close #23
